+++ Changelog.md
@@ -8,6 +8,8 @@
 - Added the registration of `promotion` entry to the Eloquent morph map
 - Added the cart item percent discount promotion action type
 - Changed the cart fixed and percent discount promotion action types to only act on carts, not on cart items
+- Added the static `getAvailableWithoutCoupon()` and `getAvailableOnes()` methods to the `Promotion` model (but not to the interface)
+- Added the `active()` and `notDepeleted()` scopes to the `Promotion` model
 
 ## 5.0.0
 ##### 2025-09-03