+++ src/Promotion/Changelog.md
@@ -7,6 +7,8 @@
 
 - Added the cart item percent discount promotion action type
 - Changed the cart fixed and percent discount promotion action types to only act on carts, not on cart items
+- Added the static `getAvailableWithoutCoupon()` and `getAvailableOnes()` methods to the `Promotion` model (but not to the interface)
+- Added the `active()` and `notDepeleted()` scopes to the `Promotion` model
 
 ## 5.0.0
 ##### 2025-09-03